requirements-dev.txt
-r requirements.txt
ruff==0.9.6
PyYAML==6.0.2
openapi-spec-validator==0.7.1
